// nexa-web/src/canvas/core/types.ts
export interface LayoutConfig {
  nodeWidth: number;
  nodeHeight: number;
  horizontalSpacing: number;
  verticalSpacing: number;
  columnWidth: number;
}